rename: comonents 파일들 분리
